+++ pyrevolut/api/payout_links/post/create_payout_link.py
@@ -1,6 +1,5 @@
 from typing import Annotated
 from uuid import UUID
-from decimal import Decimal
 
 from pydantic import BaseModel, Field
 from pydantic_extra_types.currency_code import Currency
@@ -75,7 +74,7 @@ class CreatePayoutLink:
             ),
         ]
         amount: Annotated[
-            Decimal,
+            float,
             Field(
                 description="""
                 The amount of money to be transferred.
