+++ pyrevolut/api/transactions/resources/transaction.py
@@ -1,6 +1,5 @@
 from typing import Annotated
 from uuid import UUID
-from decimal import Decimal
 
 from pydantic import BaseModel, Field
 from pydantic_extra_types.currency_code import Currency
@@ -64,11 +63,11 @@ class ResourceTransaction(BaseModel):
 
         leg_id: Annotated[UUID, Field(description="The ID of the leg.")]
         amount: Annotated[
-            Decimal,
+            float,
             Field(description="The amount of the transaction leg."),
         ]
         fee: Annotated[
-            Decimal | None,
+            float | None,
             Field(description="The amount of the transaction leg fee."),
         ] = None
         currency: Annotated[
@@ -76,7 +75,7 @@ class ResourceTransaction(BaseModel):
             Field(description="ISO 4217 currency code in upper case."),
         ]
         bill_amount: Annotated[
-            Decimal | None,
+            float | None,
             Field(description="The billing amount for cross-currency payments."),
         ] = None
         bill_currency: Annotated[
@@ -98,7 +97,7 @@ class ResourceTransaction(BaseModel):
             Field(description="The transaction leg purpose."),
         ] = None
         balance: Annotated[
-            Decimal | None,
+            float | None,
             Field(
                 description="The total balance of the account that the transaction is associated with."
             ),
